Correcciones

+++ app/Http/Controllers/AmigoController.php
@@ -33,13 +33,15 @@ class AmigoController extends Controller
     {
         //Recibe la informacion del formulario
         /* $datosAmigos = request()->all(); // Obtiene todos los campos */
-        $datosAmigos = request()->except('_token');
-
-        /* Insertar directamente a la base de datos */
-        /* Nombre del modelo / tabla, funcion o proceso */
-        Amigo::insert($datosAmigos);
-        // return response()->json($datosAmigos);
-        return redirect('amigos')->with('mensaje', 'Amigo Creado Correctamente.!');
+        if(request()->nombre){
+            $datosAmigos = request()->except('_token');
+            /* Insertar directamente a la base de datos */
+            /* Nombre del modelo / tabla, funcion o proceso */
+            Amigo::insert($datosAmigos);
+            // return response()->json($datosAmigos);
+            return redirect('amigos')->with('mensaje', 'Amigo Creado Correctamente.!');
+        }
+        return redirect('amigos');
     }
 
     /**
@@ -65,14 +67,16 @@ class AmigoController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(request()->_method === 'PUT'){
+            $datosAmigo = request()->except(['_token', '_method']);
+            Amigo::where('id', '=', $id)->update($datosAmigo);
+            /* Retorno al formulario */
+            // $amigo = Amigo::findOrFail($id);
+            // return view('amigo.edit', compact('amigo'))->with('mensaje', 'Amigo Actualizado Correctamente.!');
+            return redirect('amigos')->with('mensaje', 'Amigo Actualizado Correctamente.!');
+        }
+        return response()->json(request());
         // Actualizacion de la informacion
-        $datosAmigo = request()->except(['_token', '_method']);
-        Amigo::where('id', '=', $id)->update($datosAmigo);
-
-        /* Retorno al formulario */
-        $amigo = Amigo::findOrFail($id);
-        // return view('amigo.edit', compact('amigo'))->with('mensaje', 'Amigo Actualizado Correctamente.!');
-        return redirect('amigos')->with('mensaje', 'Amigo Actualizado Correctamente.!');
     }
 
     /**
@@ -81,7 +85,10 @@ class AmigoController extends Controller
     public function destroy($id)
     {
         //
-        Amigo::destroy($id);
-        return redirect('amigos')->with('mensaje', 'Amigo Eliminado Correctamente.!');
+        if($id){
+            Amigo::destroy($id);
+            return redirect('amigos')->with('mensaje', 'Amigo Eliminado Correctamente.!');
+        }
+        return redirect('amigos');
     }
 }
